@@ -1,11 +1,11 @@
 name: smiley_rating_dialog
 description: A new Flutter package which helps you to implement beautiful customizable Smiley Rating Dialog.
-version: 0.0.1
-author: Mohak Gupta <mohak1936@gmail.com>
+version: 1.0.0
+
 homepage: https://github.com/mohak1283/Smiley-Rating-Dialog
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.1.4 <4.0.0"
 
 dependencies:
   flutter:
@@ -51,3 +51,7 @@ flutter:
   #
   # For details regarding fonts in packages, see
   # https://flutter.io/custom-fonts/#from-packages
+
+platforms:
+  android:
+  ios:
